Add cmake, move files
