+++ .gitignore
@@ -1,8 +1,52 @@
-*.o
+# Windows image file caches
+Thumbs.db
+ehthumbs.db
+
+# Folder config file
+Desktop.ini
+
+# Recycle Bin used on file shares
+$RECYCLE.BIN/
+
+# Windows Installer files
+*.cab
+*.msi
+*.msm
+*.msp
+
+# Codeblocks files
 *.layout
 *.depend
-output/
-*.exe
-*.dll
-*.res
+
+# =========================
+# Operating System Files
+# =========================
+
+# OSX
+# =========================
+
+.DS_Store
+.AppleDouble
+.LSOverride
+
+# Icon must ends with two \r.
+Icon
+
+
+# Thumbnails
+._*
+
+# Files that might appear on external disk
+.Spotlight-V100
+.Trashes
+
 CppCheckResults.xml
+CMakeFiles
+CMakeCache.txt
+Makefile
+*.cmake
+install_manifest.txt
+bin
+docs
+lib
+build/JumpingJimothy*
