--- mazes.cbp
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6" />
-	<Project>
-		<Option title="Mazes" />
-		<Option pch_mode="2" />
-		<Option compiler="gcc" />
-		<Build>
-			<Target title="default">
-				<Option output="build/Mazes" prefix_auto="1" extension_auto="1" />
-				<Option working_dir="build" />
-				<Option object_output="output" />
-				<Option type="0" />
-				<Option compiler="gcc" />
-				<Compiler>
-					<Add option="-g" />
-				</Compiler>
-			</Target>
-		</Build>
-		<Compiler>
-			<Add option="-Wnon-virtual-dtor" />
-			<Add option="-Winit-self" />
-			<Add option="-Wredundant-decls" />
-			<Add option="-Wcast-align" />
-			<Add option="-Wundef" />
-			<Add option="-Wfloat-equal" />
-			<Add option="-Winline" />
-			<Add option="-Wunreachable-code" />
-			<Add option="-Wmissing-include-dirs" />
-			<Add option="-Wswitch-enum" />
-			<Add option="-Wswitch-default" />
-			<Add option="-Wzero-as-null-pointer-constant" />
-			<Add option="-Wmain" />
-			<Add option="-pedantic" />
-			<Add option="-Wextra" />
-			<Add option="-Wall" />
-			<Add option="-std=c++14" />
-			<Add directory="include" />
-		</Compiler>
-		<Linker>
-			<Add option="-lloadpng" />
-			<Add option="-llogg" />
-			<Add option="-lalleg44" />
-		</Linker>
-		<Unit filename="include/block.h" />
-		<Unit filename="include/button.h" />
-		<Unit filename="include/id.h" />
-		<Unit filename="res/resources.rc">
-			<Option compilerVar="WINDRES" />
-		</Unit>
-		<Unit filename="src/block.cpp" />
-		<Unit filename="src/button.cpp" />
-		<Unit filename="src/main.cpp" />
-		<Extensions>
-			<code_completion />
-			<envvars />
-			<debugger />
-		</Extensions>
-	</Project>
-</CodeBlocks_project_file>
