--- mazes.depend
@@ -1,70 +1,0 @@
-# depslib dependency file v1.0
-1396139487 source:e:\allans programming\games indev\mazes\button.cpp
-	"button.h"
-
-1396136878 e:\allans programming\games indev\mazes\button.h
-	<allegro.h>
-	<string>
-	<time.h>
-
-1398127801 source:e:\allans programming\games indev\mazes\main.cpp
-	<allegro.h>
-	<alpng.h>
-	<string>
-	<fstream>
-	<sstream>
-	<math.h>
-	<iostream>
-	"rapidxml.hpp"
-	"rapidxml_print.hpp"
-	"fmod/fmod.h"
-	"fmod/fmod_errors.h"
-	"button.h"
-	"id.h"
-
-1366484954 e:\allans programming\games indev\mazes\id.h
-
-1396112531 source:e:\allans programming\games indev\mazes\resources.rc
-
-1560647815 source:c:\users\alege\documents\github\mazes\res\resources.rc
-
-1561336769 source:c:\users\alege\documents\github\mazes\src\block.cpp
-	"block.h"
-	<loadpng.h>
-
-1561336768 c:\users\alege\documents\github\mazes\include\block.h
-	<allegro.h>
-
-1561337021 source:c:\users\alege\documents\github\mazes\src\button.cpp
-	"button.h"
-	<loadpng.h>
-
-1561337028 c:\users\alege\documents\github\mazes\include\button.h
-	<allegro.h>
-
-1561337013 source:c:\users\alege\documents\github\mazes\src\main.cpp
-	<allegro.h>
-	<loadpng.h>
-	<logg.h>
-	<string>
-	<fstream>
-	<sstream>
-	<math.h>
-	"rapidxml/rapidxml.hpp"
-	"rapidxml/rapidxml_print.hpp"
-	"button.h"
-	"id.h"
-
-1561335765 c:\users\alege\documents\github\mazes\include\rapidxml\rapidxml.hpp
-	<cstdlib>
-	<cassert>
-	<new>
-	<exception>
-
-1561335765 c:\users\alege\documents\github\mazes\include\rapidxml\rapidxml_print.hpp
-	"rapidxml/rapidxml.hpp"
-	<ostream>
-	<iterator>
-
-1560647815 c:\users\alege\documents\github\mazes\include\id.h
-
